requirements.txt
streamlit
streamlit_vtkjs
plotly
rhino3dm
compute_rhino3d
rhinoinside
honeybee-vtk
ladybug
ladybug_rhino
ladybug_geometry
honeybee
honeybee_energy
honeybee_radiance
honeybee_openstudio
numpy
pandas
matplotlib
shap
scikit-learn
joblib
openai
llama-cpp-python
tiktoken
python-docx
weasyprint
Pillow
pypandoc
odfpy
